Added a class for the shared attribute of font size for 3rd and 4th elem
Originally I had created a rule with a nested .odd.third-elem, #fourth
to change the font size for third and fourth. Later, I realized that
for the shared attributes creating a class that both elements can have
is a better way then specifying both elements as it can be more flexible
as well in the future.